fix: automatically invalidate moved/deleted entries from the ResourceForUriCache
If a resource is moved or deleted, it's possible that the resource
stored for its URI in the ResourceForUriCache stops being `accessible`.

This commit changes `ResourceForUriCache.get(URI)` so that, if a
resource is present in the cache but is no longer accessible, the
corresponding cache entry is invalidated. In this case, `.get(URI)` then
tries to determine the up-to-date IResource again.

Fixes a regression introduced in #1208. Specifically, this commit
addresses the concern raised in
https://github.com/eclipse-lsp4e/lsp4e/pull/1208#discussion_r1942357586