fix: cancel pending hover request if a new one is initiated
previously, if a hover info request could not be served in time the
method will return a dummy region and later an empty hover information.

If a new request would be then initiated, the old request would be left
running even though its result would never be consumed. To free
resources on the server side, we cancel previous request before
initiating new ones.